Change BitsAndBytes compute_dtype to float32 to prevent Gemma NaN overflows

+++ pruning/compute_fisher_masks.py
@@ -23,7 +23,7 @@ os.makedirs(output_dir, exist_ok=True)
 print(f"[*] Loading {model_id} in INT4... (Cached)")
 bnb_config = BitsAndBytesConfig(
     load_in_4bit=True,
-    bnb_4bit_compute_dtype=torch.float16,
+    bnb_4bit_compute_dtype=torch.float32,
     bnb_4bit_use_double_quant=True,
     bnb_4bit_quant_type="nf4",
     llm_int8_enable_fp32_cpu_offload=True
